refactor(web): extract webhook API router without behavior change
